Attach size and email method validate

+++ www/source/App/Web.php
@@ -471,27 +471,34 @@ class Web
                     }
                 }
 
-                $email = new Email();
-
-                $message = file_get_contents(THEMES . "/assets/emails/confirmRegisterEmail.php");
-
-                $url = ROOT . "/confirmAccount/" . md5($user->id);
-
-                $template = array("%title", "%textBody", "%button", "%link", "%name");
-                $dataReplace = array("Confirmação de Cadastro", "Para confirmar seu cadastro", "Confirmar", $url, $user->nome);
-                $message = str_replace($template, $dataReplace, $message);
-
-                $email->add(
-                    "Confirmação de cadastro",
-                    $message,
-                    $user->nome,
-                    $user->email
-                )->send();
-                if ($email->error()) {
-                    $attach->destroy();
-                    $user->destroy();
-                    var_dump($email->error()->getMessage());
-                } else {
+                $fail = false;
+                if ($user->email) {
+                    $email = new Email();
+    
+                    $message = file_get_contents(THEMES . "/assets/emails/confirmRegisterEmail.php");
+    
+                    $url = ROOT . "/confirmAccount/" . md5($user->id);
+    
+                    $template = array("%title", "%textBody", "%button", "%link", "%name");
+                    $dataReplace = array("Confirmação de Cadastro", "Para confirmar seu cadastro", "Confirmar", $url, $user->nome);
+                    $message = str_replace($template, $dataReplace, $message);
+    
+                    $email->add(
+                        "Confirmação de cadastro",
+                        $message,
+                        $user->nome,
+                        $user->email
+                    )->send();
+                    
+                    if ($email->error()) {
+                        $attach->destroy();
+                        $user->destroy();
+                        $fail = true;
+                        var_dump($email->error()->getMessage());
+                    }
+                }
+                
+                 if ($fail == false) {
                     echo "success";
                 }
             }
@@ -1747,24 +1754,29 @@ class Web
 
                 if (!$punishment->fail()) {
                     if ($license->status == 2 && $data['punishmentStatus'] == 0) {
-                        $email = new Email();
-                        $email->add(
-                            'Licença desbloqueada',
-                            'Você teve sua licença desbloqueada no orditi. Acesse seu perfil para saber mais.',
-                            $user->nome,
-                            $user->email
-                        )->send();
+                        if ($user->email) {
+                            $email = new Email();
+                            $email->add(
+                                'Licença desbloqueada',
+                                'Você teve sua licença desbloqueada no orditi. Acesse seu perfil para saber mais.',
+                                $user->nome,
+                                $user->email
+                            )->send();
+                        }
                         $license->status = 0;
                     }
 
                     if ($data['punishmentStatus'] == 1) {
-                        $email = new Email();
-                        $email->add(
-                            'Licença bloqueada',
-                            'Você teve sua licença bloqueada no orditi. Acesse seu perfil para saber mais.',
-                            $user->nome,
-                            $user->email
-                        )->send();
+                        if ($user->email) {
+                            $email = new Email();
+                            $email->add(
+                                'Licença bloqueada',
+                                'Você teve sua licença bloqueada no orditi. Acesse seu perfil para saber mais.',
+                                $user->nome,
+                                $user->email
+                            )->send();
+                        }
+                        
                         $license->status = 2;
                     }
 
@@ -1785,6 +1797,7 @@ class Web
             echo json_encode(['success' => 'Update success']);
         }
     }
+
 
     public function licenseUser($data): void
     {
@@ -2137,33 +2150,39 @@ class Web
                 exit;
             }
 
-            /**
-             * Send email with new temporary recovery password
-             */
-            $email = new Email();
-
-            $message = file_get_contents(THEMES . "/assets/emails/confirmRegisterEmail.php");
-
-            $url = ROOT . "/confirmAccount/" . md5($user->id);
-            $template = array("%title", "%textBody", "%button", "%link", "%name");
-            $dataReplace = array("Recuperação de senha", "Para recuperar sua senha", "Recuperar", $url, $user->nome);
-            $message = str_replace($template, $dataReplace, $message);
-
-            $email->add(
-                "Recuperação de senha",
-                $message,
-                $user->nome,
-                $user->email
-            )->send();
-
-            if ($email->error()) {
-                var_dump($email->error()->getMessage());
-                exit;
+            if ($user->email) {
+                /**
+                 * Send email with new temporary recovery password
+                 */
+                $email = new Email();
+    
+                $message = file_get_contents(THEMES . "/assets/emails/confirmRegisterEmail.php");
+    
+                $url = ROOT . "/confirmAccount/" . md5($user->id);
+                $template = array("%title", "%textBody", "%button", "%link", "%name");
+                $dataReplace = array("Recuperação de senha", "Para recuperar sua senha", "Recuperar", $url, $user->nome);
+                $message = str_replace($template, $dataReplace, $message);
+    
+                $email->add(
+                    "Recuperação de senha",
+                    $message,
+                    $user->nome,
+                    $user->email
+                )->send();
+    
+                if ($email->error()) {
+                    var_dump($email->error()->getMessage());
+                    exit;
+                } else {
+                    echo 'pswSuccess';
+                }                
             } else {
-                echo 'pswSuccess';
+                echo 'emailNotExist';
             }
+
         }
     }
+
 
 
     /**
@@ -3414,7 +3433,7 @@ class Web
     public function removeSuspension($data): void
     {
         $data = filter_var_array($data, FILTER_SANITIZE_STRIPPED);
-        $salesman = (new Salesman())->findById($data['id'], 'id, suspenso');
+        $salesman = (new Salesman())->findById($data['id'], 'id, id_licenca, suspenso');
         if ($salesman) {
             $payments = (new Payment())->find('id_ambulante = :id', 'id=' . $data['id'])->fetch(true);
             $aux = 0;
@@ -3433,23 +3452,29 @@ class Web
             $salesman->suspenso = 0;
             $salesman->save();
 
-            $email = new Email();
-
-            $message = file_get_contents(THEMES . "/assets/emails/removeNotificationEmail.php");
-
-            $url = "https://www.google.com";
-            $template = array("%title", "%status", "%textBody", "%button", "%link", "%name");
-            $dataReplace = array("Notificação", "MULTA", "foi removida", "Acesse", $url, $salesman->nome);
-            $message = str_replace($template, $dataReplace, $message);
-
-            $email->add(
-                "Notificação",
-                $message,
-                $salesman->nome,
-                $salesman->email
-            )->send();
-
-            echo 1;
+            $license = (new License())->findById($salesman->id_licenca);
+            if ($license) {
+                $user = (new User())->findById($license->id_usuario);
+                 if ($user->email) {
+                    $email = new Email();
+        
+                    $message = file_get_contents(THEMES . "/assets/emails/removeNotificationEmail.php");
+        
+                    $url = "https://www.google.com";
+                    $template = array("%title", "%status", "%textBody", "%button", "%link", "%name");
+                    $dataReplace = array("Notificação", "MULTA", "foi removida", "Acesse", $url, $salesman->nome);
+                    $message = str_replace($template, $dataReplace, $message);
+        
+                    $email->add(
+                        "Notificação",
+                        $message,
+                        $salesman->nome,
+                        $salesman->email
+                    )->send();
+                }   
+                
+                echo 1;
+            }
         } else {
             echo 0;
         }
